+++ e2e/fixtures/view-transition/rspress.config.ts
@@ -1,0 +1,9 @@
+import * as path from 'path';
+import { defineConfig } from 'rspress/config';
+
+export default defineConfig({
+  themeConfig: {
+    enableContentAnimation: true,
+  },
+  root: path.join(__dirname, 'doc'),
+});