+++ packages/core/src/theme-default/index.ts
@@ -22,3 +22,5 @@ export default {
 export { getCustomMDXComponent } from './layout/DocLayout/docComponents';
 
 export * from './logic';
+
+export { Layout, NotFoundLayout, HomeLayout };