+++ packages/core/src/theme-default/logic/index.ts
@@ -3,4 +3,5 @@ export { useEditLink } from './useEditLink';
 export { useSidebarData } from './useSidebarData';
 export { useLocaleSiteData } from './useLocaleSiteData';
 export { setup, bindingAsideScroll, scrollToTarget } from './sideEffects';
+export { usePathUtils } from './usePathUtils';
 export * from './utils';