+++ packages/modern-plugin-rspress/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @modern-js/plugin-rspress
 
+## 0.2.0
+
+### Minor Changes
+
+- da74868: using view transition to present animation when page navigate
+
+### Patch Changes
+
+- Updated dependencies [100e0f6]
+- Updated dependencies [b56c083]
+- Updated dependencies [178ba42]
+- Updated dependencies [da74868]
+  - @rspress/plugin-preview@0.2.0
+  - @rspress/core@0.2.0
+  - @rspress/plugin-api-docgen@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes